% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureDeleteContainer.R
\name{azureDeleteContainer}
\alias{azureDeleteContainer}
\title{A function to delete a Blob  container}
\usage{
azureDeleteContainer(sa, container, key, verbose = FALSE)
}
\arguments{
\item{sa}{"Azure Storage Account"}

\item{container}{"Name of the container"}

\item{key}{"SAS key to access the blob"}

\item{verbose(optional)}{"set to TRUE to get full information about the REST call"}
}
\description{
A function to delete a Blob  container
}

