% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azurePutBlob.R
\name{azurePutBlob}
\alias{azurePutBlob}
\title{A function to save a data structure into a  Blob  container.}
\usage{
azurePutBlob(sa, container, key, file, data, verbose = FALSE)
}
\arguments{
\item{sa}{"Azure Storage Account"}

\item{container}{"Name of the container"}

\item{key}{"SAS key to access the blob"}

\item{file}{"Filename"}

\item{data}{"Datastructure to be saved"}

\item{verbose(optional)}{"set to TRUE to get full information about the REST call"}
}
\description{
A function to save a data structure into a  Blob  container.
}

