% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureCreateContainer.R
\name{azureCreateContainer}
\alias{azureCreateContainer}
\title{A function to create a Blob  container}
\usage{
azureCreateContainer(sa, container, key)
}
\arguments{
\item{sa}{"Azure Storage Account"}

\item{container}{"Name of the container"}

\item{key}{"SAS ket to access the blob"}
}
\description{
A function to create a Blob  container
}

