% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureDeleteContainer.R
\name{azureDeleteContainer}
\alias{azureDeleteContainer}
\title{A function to create a Blob  container}
\usage{
azureDeleteContainer(sa, container, key)
}
\arguments{
\item{sa}{"Azure Storage Account"}

\item{container}{"Name of the container"}

\item{key}{"SAS key to access the blob"
azureDeleteContainer()}
}
\description{
A function to create a Blob  container
}

