% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureGetBlob.R
\name{azureGetBlob}
\alias{azureGetBlob}
\title{A function to get a file from a  Blob  container and store it in memory.}
\usage{
azureGetBlob(sa, container, key, file, verbose = FALSE)
}
\arguments{
\item{sa}{"Azure Storage Account"}

\item{container}{"Name of the container"}

\item{key}{"SAS key to access the blob"}

\item{file}{"File to upload"}

\item{verbose(optional)}{"set to TRUE to get full information about the REST call"}
}
\description{
A function to get a file from a  Blob  container and store it in memory.
}

