% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureUploadBlob.R
\name{azureUploadBlob}
\alias{azureUploadBlob}
\title{A function to upload a file to a Blob  container}
\usage{
azureUploadBlob(sa, container, key, file)
}
\arguments{
\item{sa}{"Azure Storage Account"}

\item{container}{"Name of the container"}

\item{key}{"SAS ket to access the blob"}

\item{file}{"File to upload"
azureUploadBlob()}
}
\description{
A function to upload a file to a Blob  container
}

